Address Copilot review: split files, throw on invalid protocol, nullable TryParse
- Split AdbPortSpec into its own file (one public type per file convention)
- AdbPortRule.cs now contains only AdbPortRule record
- ToSocketSpec() throws ArgumentOutOfRangeException for unknown AdbProtocol values
- TryParse parameter changed from string to string? to match null-handling behavior
- Remove unused 'using System' from AdbPortRule.cs
- Update PublicAPI files for TryParse signature change
