+++ src/Xamarin.Android.Tools.AndroidSdk/Runners/AdbPortRule.cs
@@ -1,58 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-
 namespace Xamarin.Android.Tools;
-
-/// <summary>
-/// Represents a port and protocol pair for adb forwarding/reverse operations.
-/// </summary>
-public record AdbPortSpec (AdbProtocol Protocol, int Port)
-{
-	/// <summary>
-	/// Returns the adb socket spec string, e.g. "tcp:5000".
-	/// </summary>
-	public string ToSocketSpec () => Protocol switch {
-		AdbProtocol.Tcp => $"tcp:{Port}",
-		AdbProtocol.LocalAbstract => $"localabstract:{Port}",
-		AdbProtocol.LocalReserved => $"localreserved:{Port}",
-		AdbProtocol.LocalFilesystem => $"localfilesystem:{Port}",
-		_ => $"tcp:{Port}",
-	};
-
-	/// <summary>
-	/// Parses an adb socket spec string like "tcp:5000" into an <see cref="AdbPortSpec"/>.
-	/// Returns null if the format is unrecognized.
-	/// </summary>
-	public static AdbPortSpec? TryParse (string socketSpec)
-	{
-		if (string.IsNullOrWhiteSpace (socketSpec))
-			return null;
-
-		var colonIndex = socketSpec.IndexOf (':');
-		if (colonIndex <= 0 || colonIndex >= socketSpec.Length - 1)
-			return null;
-
-		var protocolStr = socketSpec.Substring (0, colonIndex);
-		var portStr = socketSpec.Substring (colonIndex + 1);
-
-		if (!int.TryParse (portStr, out var port) || port <= 0 || port > 65535)
-			return null;
-
-		var protocol = protocolStr.ToLowerInvariant () switch {
-			"tcp" => (AdbProtocol?) AdbProtocol.Tcp,
-			"localabstract" => AdbProtocol.LocalAbstract,
-			"localreserved" => AdbProtocol.LocalReserved,
-			"localfilesystem" => AdbProtocol.LocalFilesystem,
-			_ => null,
-		};
-
-		return protocol.HasValue ? new AdbPortSpec (protocol.Value, port) : null;
-	}
-
-	public override string ToString () => ToSocketSpec ();
-}
 
 /// <summary>
 /// Represents an adb port forwarding rule as reported by 'adb reverse --list' or 'adb forward --list'.
